+++ src/hiero_sdk_python/__init__.py
@@ -111,6 +111,9 @@ from .contract.ethereum_transaction import EthereumTransaction
 from .schedule.schedule_create_transaction import ScheduleCreateTransaction
 from .schedule.schedule_id import ScheduleId
 
+# Nodes
+from .nodes.node_create_transaction import NodeCreateTransaction
+
 __all__ = [
     # Client
     "Client",
@@ -220,4 +223,7 @@ __all__ = [
     # Schedule
     "ScheduleCreateTransaction",
     "ScheduleId",
+
+    # Nodes
+    "NodeCreateTransaction",
 ]